back/requirements.txt
dj-database-url==0.5.0
Django==4.0.8
psycopg2-binary==2.9.3
wheel==0.38.1
whitenoise==6.0.0
djangorestframework
django-cors-headers
python-decouple==3.8
Pillow
djangorestframework-simplejwt
django-extensions==3.2.1
stripe==5.4.0
whitenoise
gunicorn
psycopg2==2.9.6
python-dotenv==1.0.0

